Refactor: migrate away from PhysAddr, VirtAddr, and u64
We use usize in most of our own APIs, but the x86_64 crates uses u64
everywhere, including in its PhysAddr and VirtAddr.
This commit encapsulates the PhysAddr and VirtAddr inside the mmu/frames
module so that we uniformly expose HostPhysAddr, HostVirtAddr and usize.

+++ src/guests/identity.rs
@@ -152,8 +152,8 @@ fn setup_ept(allocator: &impl FrameAllocator) -> Result<EptMapper, ()> {
 
     ept_mapper.map_range(
         allocator,
-        vmx::GuestPhysAddr::new(start as usize),
-        vmx::HostPhysAddr::new(start as usize),
+        vmx::GuestPhysAddr::new(start),
+        vmx::HostPhysAddr::new(start),
         end as usize,
         EptEntryFlags::READ | EptEntryFlags::WRITE | EptEntryFlags::SUPERVISOR_EXECUTE,
     );
